% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countKmer.R
\name{calculateCounts}
\alias{calculateCounts}
\title{Calculate k-mer counts}
\usage{
calculateCounts(study_table, k)
}
\arguments{
\item{study_table}{A tibble consisting of antigen receptor
sequencing imported by the LymphoSeq2 function \code{\link[=readImmunoSeq]{readImmunoSeq()}}.
"repertoire_id" and "junction" are required columns.}

\item{k}{The length of k-mers to find.}
}
\description{
Calculate k-mer counts
}
