% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonalRelatedness.R
\name{clonalRelatedness}
\alias{clonalRelatedness}
\title{Clonal relatedness}
\usage{
clonalRelatedness(study_table, edit_distance = 10)
}
\arguments{
\item{study_table}{A tibble of raw or productive nucleotide junction sequences.
"junction" and "duplicate_count" are required columns.}

\item{edit_distance}{An integer giving the minimum edit distance that the
sequence must be less than or equal to. See details below. Deafult 10 (the default)}
}
\value{
Returns a tibble with the calculated clonal relatedness for each
repertoire_id.
}
\description{
Calculates the clonal relatedness for each repertoire in a study.
}
\details{
Clonal relatedness is the proportion of junction sequences that
are related by a defined edit distance threshold.  The value ranges from 0 to
1 where 0 indicates no sequences are related and 1 indicates all sequences
are related.

Edit distance is a way of quantifying how dissimilar two sequences
are to one another by counting the minimum number of operations required to
transform one sequence into the other. For example, an edit distance of 0
means the sequences are identical and an edit distance of 1 indicates that
the sequences differ by a single amino acid or junction.
}
\examples{
file_path <- system.file("extdata", "IGH_sequencing", package = "LymphoSeq2")
study_table <- LymphoSeq2::readImmunoSeq(path = file_path, threads = 1) 
study_table <- LymphoSeq2::topSeqs(study_table, top = 100)
clonal_relatedness <- LymphoSeq2::clonalRelatedness(study_table, edit_distance = 10)
# Merge results with clonality table
clonality <- LymphoSeq2::clonality(study_table)
merged <- dplyr::full_join(clonality, clonal_relatedness, by = "repertoire_id")

}
