% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneTrack.R
\name{cloneTrack}
\alias{cloneTrack}
\title{Clone tracking line plot}
\usage{
cloneTrack(study_table, sample_list = NULL, sequence_track = NULL)
}
\arguments{
\item{study_table}{A tibble of productive amino acid sequences
generated by LymphoSeq2 function productiveSeq where the aggregate parameter
was set to "junction_aa"}

\item{sample_list}{A character vector of one or more repertoire_ids to track.
If set to NULL (default), all repertoire_ids in the sequence matrix will be tracked.}

\item{sequence_track}{An optional character vector of one or more amino acid
sequences to track. If set to NULL (default), will pull all junction_aa sequences
from the sequence matrix.}
}
\value{
Returns a line plot showing the amino acid frequencies across
multiple samples in the sequence matrix where each line represents one
unique sequence.
}
\description{
Creates line plot tracking amino acid frequencies across multiple samples
}
\details{
The plot is made using the package ggplot2 and can be reformatted
using ggplot2 functions.  See examples below.
}
\examples{
file_path <- system.file("extdata", "TCRB_sequencing", package = "LymphoSeq2")

stable <- readImmunoSeq(path = file_path)

atable <- productiveSeq(study_table = stable, aggregate = "junction_aa")

top_freq <- topFreq(atable, frequency = 0.001)

# Track clones without mapping or tracking specific sequences
cloneTrack(atable)

# Track top 20 clones mapping to the CD4 and CD8 samples
cloneTrack(atable, sample_list = c("TRB_CD4_949", "TRB_CD8_949"),
   sequence_track = top_freq$junction_aa[1:20]) 

# Track the top 10 clones from top.freq
cloneTrack(study_table = atable, sequence_track = top_freq$junction_aa[1:10]) 

# Track clones mapping to the CD4 and CD8 samples while ignoring all others
cloneTrack(study_table = atable, sample_list = c("TRB_CD4_949", "TRB_CD8_949"))

# Track clones mapping to the CD4 and CD8 samples and track 2 specific sequences
cloneTrack(study_table = atable, sample_list = c("TRB_CD4_949", "TRB_CD8_949"),
   sequence_track = c("CASSPPTGERDTQYF", "CASSQDRTGQYGYTF"))

}
\seealso{
An excellent resource for examples on how to reformat a ggplot can
be found in the R Graphics Cookbook online (\url{http://www.cookbook-r.com/Graphs/}).
}
