% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read10x.R
\name{collapse_chains}
\alias{collapse_chains}
\title{Read 10x Genomics data containing alpha and beta chains}
\usage{
collapse_chains(clone_file)
}
\arguments{
\item{clone_file}{A single .tsv file from 10x Genomics}
}
\value{
A tibble with alpha and beta chains collapsed
}
\description{
Read in 10x Genomics data and collapse alpha and beta
chains appropriately. This is for legacy 10x formats.
}
