% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonSeqs.R
\name{commonSeqs}
\alias{commonSeqs}
\title{Common sequences in two or more repertoire_ids}
\usage{
commonSeqs(study_table, repertoire_ids = NULL)
}
\arguments{
\item{study_table}{A list of productive amino acid sequences generated
by the LymphoSeq2 function productiveSeq where aggregate = "junction_aa".}

\item{repertoire_ids}{A character vector of two or more repertoire_id names in
study_table.}
}
\value{
Returns a data frame of the common sequences between two or more files
displaying their frequencies in each.
}
\description{
Creates a data frame of the common sequences in two or more repertoire_ids, reporting
their frequencies in each.
}
\examples{
file_path <- system.file("extdata", "TCRB_sequencing", package = "LymphoSeq2")

stable <- readImmunoSeq(path = file_path)

atable <- productiveSeq(study_table = stable, aggregate = "junction_aa")

commonSeqs(repertoire_ids = c("TRB_Unsorted_0", "TRB_Unsorted_32"),
   study_table = atable)
}
\seealso{
\code{\link{productiveSeq}} \code{\link{commonSeqsVenn}} \code{\link{commonSeqsPlot}} \code{\link{commonSeqsBar}}
}
