% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonSeqsBar.R
\name{commonSeqsBar}
\alias{commonSeqsBar}
\title{Common sequences bar plot}
\usage{
commonSeqsBar(
  amino_table,
  repertoire_ids,
  color_sample = NULL,
  color_intersection = NULL,
  color = "#377eb8",
  labels = "no"
)
}
\arguments{
\item{amino_table}{A tibble of productive amino acid sequences
generated by LymphoSeq2 function productiveSeq where the aggregate parameter
was set to "junction_aa".}

\item{repertoire_ids}{The names of two or more repertoire_ids in the
amino_table list whose intersections will shown.}

\item{color_sample}{The name of a single repertoire_id in the amino_table
whose sequences will be colored in all repertoire_ids that they appear in.}

\item{color_intersection}{The names of two or more repertoire_ids in the
amino_table intersections will be colored.}

\item{color}{A character vector of a color name that will be used highlight
a selected repertoire_id or multiple repertoire_id intersections.}

\item{labels}{A character vector indicating whether the number of
intersecting sequences should be shown on the tops of the bars.Options
include "yes" or "no".}
}
\value{
Returns an UpSetR plot showing the number of
intersecting sequences across multiple repertoire_ids.
}
\description{
Creates an UpSet plot showing the number of intersecting sequences
across multiple repertoire_ids using the UpSetR package. This
function is useful when more than 3 repertoire_ids are being compared.
}
\examples{
\dontrun{
file_path <- system.file("extdata", "TCRB_sequencing",
 package = "LymphoSeq2")
study_table <- LymphoSeq2::readImmunoSeq(path = file_path, threads = 1)
study_table <- LymphoSeq2::topSeqs(study_table, top = 100)
amino_table <- LymphoSeq2::productiveSeq(study_table,
 aggregate = "junction_aa")
LymphoSeq2::commonSeqsBar(amino_table, repertoire_ids = c(
  "TRB_CD4_949", "TRB_CD8_949",
  "TRB_Unsorted_949", "TRB_Unsorted_1320"
), color_sample = "TRB_CD8_949")
}
}
\seealso{
\code{\link[=productiveSeq]{productiveSeq()}}, \code{\link[=commonSeqs]{commonSeqs()}},
\code{\link[=commonSeqsVenn]{commonSeqsVenn()}}, \code{\link[=commonSeqsPlot]{commonSeqsPlot()}}
}
