% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonSeqsPlot.R
\name{commonSeqsPlot}
\alias{commonSeqsPlot}
\title{Common sequences scatter plot}
\usage{
commonSeqsPlot(sample1, sample2, productive_aa, show = "common")
}
\arguments{
\item{sample1}{A name of a repertoire_id in a list of data frames generated by the
LymphoSeq2 function productiveSeq.}

\item{sample2}{A name of a repertoire_id in a list of data frames generated by the
LymphoSeq function productiveSeq.}

\item{productive_aa}{A tibble of productive amino acid sequences
produced by the LymphoSeq function productiveSeq containing the
samples to be compared.}

\item{show}{A character vector specifying whether only the common sequences
should be shown or all sequences.  Available options are "common" or "all".}
}
\value{
Returns a frequency scatter plot of two samples showing only the
shared sequences.
}
\description{
Creates a scatter plot of just the sequences in common between two samples.
}
\details{
The plot is made using the package ggplot2 and can be reformatted
using ggplot2 functions. See examples below.
}
\examples{
file_path <- system.file("extdata", "TCRB_sequencing", package = "LymphoSeq2")

stable <- readImmunoSeq(path = file_path)

atable <- productiveSeq(study_table = stable, aggregate = "junction_aa")

commonSeqsPlot("TRB_Unsorted_32", "TRB_Unsorted_83", 
   productive_aa = atable)

# Change the X and Y axis to log-10 scale
commonSeqsPlot("TRB_Unsorted_32", "TRB_Unsorted_83", 
   productive_aa = atable) +
   ggplot2::scale_x_log10() + 
   ggplot2::scale_y_log10() + 
   ggplot2::annotation_logticks(sides = "bl")
}
\seealso{
An excellent resource for examples on how to reformat a ggplot can
be found in the R Graphics Cookbook online (\url{http://www.cookbook-r.com/Graphs/}).

\code{\link{productiveSeq}} \code{\link{commonSeqsVenn}} \code{\link{commonSeqsPlot}} \code{\link{commonSeqsBar}}
}
