% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialAbundance.R
\name{differentialAbundance}
\alias{differentialAbundance}
\title{Differential abundance analysis}
\usage{
differentialAbundance(
  study_table,
  repertoire_ids = NULL,
  abundance = "duplicate_count",
  type = "junction_aa",
  q = 1,
  zero = 1,
  parallel = FALSE
)
}
\arguments{
\item{abundance}{The input value for the Fisher exact test.  "duplicate_count"
is recommend but "duplicate_count" may also be used.}

\item{type}{A character vector indicating whether "junction_aa" or "junction" sequences
should be used.  If "junction_aa" is specified, then run productiveSeqs first.}

\item{q}{A numeric value between 0.0 and 1.0 indicating the threshold Holms adjusted 
P value (also knowns as the false discovery rate or q value) to subset the results with.  
Any sequences with a q value greater than this value will not be shown.}

\item{zero}{A numeric value to set all zero values to when calculating the log2
transformed fold change between samples 1 and 2.  This does not apply to the 
p and q value calculations.}

\item{parallel}{A boolean indicating wheter parallel processing should be used or not.}

\item{list}{A tibble consisting of antigen receptor sequences 
imported by the LymphoSeq function readImmunoSeq.}

\item{sample1}{A character vector indicating the name of the first repertoire_id in the list
to be compared.}

\item{sample2}{A character vector indicating the name of the second repertoire_id in the list
to be compared.}
}
\value{
Returns a data frame with columns corresponding to the frequency of the abudance
measure in samples 1 and 2, the P value, Q value (Holms adjusted P value, also knowns as 
the false discovery rate), and log2 transformed fold change.
}
\description{
Use a Fisher exact test to calculate differential abdunance of each sequence in 
two samples and reports the log2 transformed fold change, P value and adjusted P value.
}
\examples{
file.path <- system.file("extdata", "TCRB_sequencing", package = "LymphoSeq2")

study_table <- readImmunoSeq(path = file.path)

productive_aa <- productiveSeq(study_table = study_table, aggregate = "junction_aa")

differentialAbundance(study_table = productive_aa, repertoire_ids = c("TRB_Unsorted_949", "TRB_Unsorted_1320"), 
                      type = "junction_aa", q = 0.01, zero = 0.001)
}
