% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonSeqsVenn.R
\name{draw_venn3_native}
\alias{draw_venn3_native}
\title{Draw 3-way Venn diagram with ggplot2}
\usage{
draw_venn3_native(seqs_list, labels)
}
\arguments{
\item{seqs_list}{Named list of sequence vectors}

\item{labels}{Character vector of set names}
}
\value{
ggplot object
}
\description{
Draw 3-way Venn diagram with ggplot2
}
\keyword{internal}
