% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runINext.R
\name{extrapolation_diversity}
\alias{extrapolation_diversity}
\title{Calculate extrapolation diversity}
\usage{
extrapolation_diversity(abundance, m, q)
}
\arguments{
\item{abundance}{Vector of species abundances}

\item{m}{Sample size (> observed)}

\item{q}{Diversity order}
}
\value{
Diversity estimate
}
\description{
Calculate extrapolation diversity
}
\keyword{internal}
