% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialAbundance.R
\name{fisherFunction}
\alias{fisherFunction}
\title{Calculates}
\usage{
fisherFunction(x, y, not_x, not_y)
}
\arguments{
\item{x}{duplicate count of first repertoire_id}

\item{y}{duplicate count of second repertoire_id}

\item{not_x}{duplicate count of values not equal to first repertoire_id}

\item{not_y}{duplicate count of values not equal to second repertoire_id}
}
\value{
list of calculated values
}
\description{
Calculates fisher test, p, and logorithm values
}
