% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneFreq.R
\name{geneFreq}
\alias{geneFreq}
\title{Gene frequencies}
\usage{
geneFreq(nucleotide_table, locus = "VDJ", family = FALSE)
}
\arguments{
\item{nucleotide_table}{A tibble of productive sequences
generated by the LymphoSeq2 function \code{\link[=productiveSeq]{productiveSeq()}} where the parameter
aggregate is set to "junction".}

\item{locus}{A character vector indicating which VDJ genes to include in the
output.  Available options include \code{"VDJ"}, \code{"DJ"}, \code{"VJ"}, \code{"DJ"}, \code{"V"},
\code{"D"}, or \code{"J"}.}

\item{family}{A Boolean value indicating whether or not family names instead
of gene names are used.  If TRUE, then family names are used and if FALSE,
gene names are used.}
}
\value{
Returns a data frame with the repertoire_id names, VDJ gene name, duplicate_count, and
\\% frequency of the V, D, or J genes (each gene frequency should add to
100\\% for each repertoire_id).
}
\description{
Creates a data frame of VDJ gene counts and frequencies.
}
\examples{
file_path <- system.file("extdata", "TCRB_sequencing", package = "LymphoSeq2")
study_table <- LymphoSeq2::readImmunoSeq(path = file_path) 
study_table <- LymphoSeq2::topSeqs(study_table, top = 100)
nucleotide_table <- LymphoSeq2::productiveSeq(study_table = study_table, 
  aggregate = "junction")
LymphoSeq2::geneFreq(nucleotide_table, locus = "VDJ", family = FALSE)
}
