% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readImmunoSeq.R
\name{getFileType}
\alias{getFileType}
\title{Identify file type:}
\usage{
getFileType(clone_file)
}
\arguments{
\item{clone_file}{.tsv file containing results from AIRRSeq pipeline}
}
\value{
List containing file type and standardized header names
}
\description{
AIRRSeq results have various formats, this functions identifies six formats
}
