% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGliphTable.R
\name{getGliphTable}
\alias{getGliphTable}
\title{Group productive sequences by repertoire}
\usage{
getGliphTable(gliph_path, progress_bar)
}
\arguments{
\item{gliph_path}{Path to individual GLIPH file}

\item{progress_bar}{Progress bar variable}
}
\description{
Group productive sequences by repertoire
}
