% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getPrevalenceTRB}
\alias{getPrevalenceTRB}
\title{Prevalence TRB}
\source{
TCR beta sequencing data from PBMCs of 55 healthy individuals.
See: https://bioconductor.org/packages/LymphoSeqDB
}
\usage{
getPrevalenceTRB()
}
\value{
A tibble with 11,724,294 rows and 2 columns (prevalence, aminoAcid)
}
\description{
Access prevalence data from the LymphoSeqDB package.
The database describes the frequency at which a CDR3 amino acid sequence was
found in cohort of 55 PBMC samples from healthy individuals.
}
\details{
This function loads data from the LymphoSeqDB Bioconductor package.
Install with: BiocManager::install("LymphoSeqDB")
}
\examples{
\dontrun{
prevalenceTRB <- getPrevalenceTRB()
}
}
