% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readImmunoSeq.R
\name{getStandard}
\alias{getStandard}
\title{Get standard headers}
\usage{
getStandard(file_type)
}
\arguments{
\item{file_type}{file type from getFileType method}
}
\value{
List of standard header names
}
\description{
Get standard headers
}
