% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countKmer.R
\name{kmerPlot}
\alias{kmerPlot}
\title{Plot kmer distributions}
\usage{
kmerPlot(kmer_table, top = 10)
}
\arguments{
\item{kmer_table}{A tibble of k-mer counts generated by the LymphoSeq2
function countKmer where the separate parameter is set to TRUE.}

\item{top}{The number of top k-mer to show}
}
\value{
A stacked bar chart showing k-mer distributions by repertoire_id
}
\description{
Plot k-mer distributions by repertoire id
}
\examples{
file_path <- system.file("extdata", "TCRB_sequencing",
 package = "LymphoSeq2")
study_table <- LymphoSeq2::readImmunoSeq(path = file_path, threads = 1)
study_table <- LymphoSeq2::topSeqs(study_table, top = 100)
kmer_table <- LymphoSeq2::countKmer(study_table = study_table, k = 5,
 separate = TRUE)
kmer_distributions <- LymphoSeq2::kmerPlot(kmer_table, top = 10)
}
