% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read10x.R
\name{mostPrevalent}
\alias{mostPrevalent}
\title{Selecting the alpha and beta chains}
\usage{
mostPrevalent(barcode_data, clone_data)
}
\arguments{
\item{barcode_data}{A tibble that holds data for one barcode identifier}

\item{clone_data}{A tibble that holds data for one repertoire_id (one file)}
}
\value{
A tibble containing the most frequent alpha chain and
the most frequent beta chain.
}
\description{
Select the most frequently occurring alpha and beta chains
for each barcode
}
