% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPairwiseNetwork.R
\name{pairwiseNetwork}
\alias{pairwiseNetwork}
\title{Pairwise comparison network graph}
\usage{
pairwiseNetwork(matrix)
}
\arguments{
\item{matrix}{A Bhattacharyya, Similarity, Sorensen, or PSI matrix produced
by the LymphoSeq2 scoringMatrix function.}
}
\value{
A network graph visualizing pairwise comparisons. The thicker the
line connecting two nodes, the greater the similarity.
}
\description{
Creates a network graph from a Bhattacharyya, Similarity, Sorensen, or PSI matrix.
}
\examples{

file_path <- system.file("extdata", "TCRB_sequencing", package = "LymphoSeq2")
stable <- LymphoSeq2::readImmunoSeq(path = file_path)
atable <- LymphoSeq2::productiveSeq(study_table = stable, aggregate = "junction_aa")
matrix <- LymphoSeq2::scoringMatrix(atable, mode = "Similarity")
network_graph <- LymphoSeq2::pairwiseNetwork(matrix)

}
