% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoringMatrix.R
\name{percentSI}
\alias{percentSI}
\title{Percent similarity index}
\usage{
percentSI(sample_list)
}
\arguments{
\item{sample_list}{A list of two tibbles consisting of frequencies of
antigen receptor sequences derived from the productiveSeq function
in LymphoSeq2.  "junction_aa", "duplicate_count" and 'repertoire_id'
are required columns.}
}
\value{
Returns the similarity score, a measure of the amount of
overlap between two samples.  The value ranges from 0 to 1 where 1 indicates
the sequence frequencies are identical in the two samples and 0
indicates no shared frequencies.
}
\description{
Calculates the Percent similarity index between two groups of repertoires. Percent
similarity index, not only compares the number of similar and dissimilar species
present between two sites, but also incorporate abundance.
}
\examples{
file_path <- system.file("extdata", "TCRB_sequencing", package = "LymphoSeq2")

stable <- readImmunoSeq(path = file_path, threads = 1)

atable <- productiveSeq(stable, aggregate = "junction_aa")

}
\seealso{
\code{\link{scoringMatrix}}
}
