% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefactionCurve.R
\name{plotRarefactionCurve}
\alias{plotRarefactionCurve}
\title{Plot rarefaction and extrapolation curves for samples}
\usage{
plotRarefactionCurve(study_table, endpoint = 1e+05)
}
\arguments{
\item{study_table}{A tibble consisting antigen receptor sequencing
data imported by the LymphoSeq2 function \code{\link[=readImmunoSeq]{readImmunoSeq()}}. "junction_aa",
"duplicate_count", and "duplicate_frequency" are required columns.}
}
\description{
Given a study table, for each sample plot rarefaction curves to estimate
repertoire diversity. The method used to generate the rarefaction curve
is derived from Chao et al., (2014) using the iNEXT library
}
\examples{
file_path <- system.file("extdata", "TCRB_sequencing",
 package = "LymphoSeq2")
study_table <- LymphoSeq2::readImmunoSeq(path = file_path, threads = 1)
study_table <- LymphoSeq2::topSeqs(study_table, top = 100)
LymphoSeq2::plotRarefactionCurve(study_table)

}
\seealso{
\code{\link[=runINext]{runINext()}}
}
