% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrack.R
\name{plotTrack}
\alias{plotTrack}
\title{Plot clone}
\usage{
plotTrack(clone_table, alist = NULL, apal = NULL)
}
\arguments{
\item{clone_table}{A tibble of productive amino acid sequences 
generated by LymphoSeq function cloneTrack}

\item{alist}{An optional list of amino acid, if a list is provided only those sequences will be highlighted}
}
\value{
An alluvial diagram tracking particular clone across samples.
}
\description{
Creates alluvial tracking amino acid frequencies across multiple samples
}
\details{
The plot is made using the package ggplot2 and can be reformatted
using ggplot2 functions.  See examples below.
}
