% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrackSingular.R
\name{plotTrackSingular}
\alias{plotTrackSingular}
\alias{highlightPlot}
\title{Highlight clones}
\usage{
plotTrackSingular(ctable)

highlightPlot(aseq, ctable)
}
\arguments{
\item{ctable}{tibble of productive amino acid sequences to highlight
generated by LymphoSeq function cloneTrack}

\item{aseq}{CDR3 amino acid sequence to highlight}

\item{clone_table}{A tibble of productive amino acid sequences to highlight
generated by LymphoSeq function cloneTrack}
}
\value{
A list of alluvial plots highlighting a single sequence
}
\description{
Create alluvial plots highlighting each sequence in amino acid list
}
\details{
The plot is made using the package ggplot2 and can be reformatted
using ggplot2 functions.
}
\section{Functions}{
\itemize{
\item \code{highlightPlot}: Highligh a specific amino acid sequence
}}

