% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{publishedTRB}
\alias{publishedTRB}
\title{Published TRB}
\format{
\subsection{\code{publishedTRB}}{

A tibble with 11,724,294 rows and 2 columns:
\describe{
\item{prevalence}{Frequency of sequences in 55 healthy PBMC samples}
\item{aminoAcid}{T-cell receptor beta chain amino acid sequence}
\item{PMID}{Pubmed ID}
\item{HLA}{Known MHC restriction}
\item{antigen}{Antigen name}
\item{epitope}{Known epitope sequence}
}
}
}
\source{
T-cells previously published T-cell beta chain CDR3 amino acid sequences
with known antigenic specificity
}
\usage{
publishedTRB
}
\description{
The database describes T-cell beta chain amino acid sequences found to be
associated with antigenic specificity
}
\keyword{datasets}
