% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runINext.R
\name{rarefaction_diversity}
\alias{rarefaction_diversity}
\title{Calculate rarefaction diversity (interpolation)}
\usage{
rarefaction_diversity(abundance, m, q)
}
\arguments{
\item{abundance}{Vector of species abundances}

\item{m}{Sample size}

\item{q}{Diversity order}
}
\value{
Diversity estimate
}
\description{
Calculate rarefaction diversity (interpolation)
}
\keyword{internal}
