% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read10x.R
\name{read_airr_format}
\alias{read_airr_format}
\title{Read AIRR-formatted 10x file}
\usage{
read_airr_format(file_path)
}
\arguments{
\item{file_path}{Path to AIRR-formatted .tsv file}
}
\value{
Tibble with AIRR data
}
\description{
Read AIRR-formatted 10x file
}
