% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read10x.R
\name{read_contig_annotations}
\alias{read_contig_annotations}
\title{Read contig_annotations format}
\usage{
read_contig_annotations(file_path)
}
\arguments{
\item{file_path}{Path to contig_annotations.csv file}
}
\value{
Tibble mapped to AIRR format
}
\description{
Read contig_annotations format
}
