% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readImmunoSeq.R
\name{readFiles}
\alias{readFiles}
\title{Read clone file from path}
\usage{
readFiles(clone_file, progress)
}
\arguments{
\item{clone_file}{.tsv file containing results from AIRRSeq pipeline}
}
\value{
Tibble in standardized format
}
\description{
Read clone file from path
}
