% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGliphTable.R
\name{readGliph}
\alias{readGliph}
\title{Read GLIPH files}
\usage{
readGliph(gliph_path)
}
\arguments{
\item{gliph_path}{Path containing GLIPH convergence group files. The file name will be used 
as repertoire_id in the output, it is advisable to rename the GLIPH files to match input 
TRB file names so that the GLIPH table and repertoire table can be merged}
}
\value{
Tibble with four columns, repertoire_id, gliph count, specificity group and
junction amino acid sequence. The GLIPH file name is used as the repertoire_id
}
\description{
\code{readGliph} Imports tab-separated value (.tsv) files generated by GLIPH
}
\details{
#' @details The files contain three columns, the gliph convergence group count,
The specificity group sequence and a space separated list of junction_aa sequences.
The function reads these files into a tibble with three columns, expanding
the junction_aa column such that each row in the tibble corresponds to one junction_aa
}
\examples{
file_path <- base::system.file("extdata", "TCRB_gliph", package = "LymphoSeq2") 
gliph_table <- LymphoSeq2::readGliph(file_path)

}
