% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqMatrix.R
\name{seqMatrix}
\alias{seqMatrix}
\title{Sequence matrix}
\usage{
seqMatrix(productive_aa, sequences = NULL, by = "duplicate_frequency")
}
\arguments{
\item{productive_aa}{A tibble of productive amino acid sequences
generated by LymphoSeq function productiveSeq where the aggregate parameter
was set to "junction_aa".}

\item{sequences}{A character vector of amino acid sequences of interest.  It
is useful to specify the output from the LymphoSeq functions uniqueSeqs or
topSeqs and subsetting the "junction_aa" column.  See examples below.}

\item{by}{Available options are "duplicate_frequency" and "duplicate_count".
Default is "duplicate_frequence".}
}
\value{
Returns a data frame of unique, productive amino acid sequences as
rows and the \\% frequency it appears in each repertoire_id as columns.
}
\description{
Creates a data frame with unique, productive amino acid sequences as rows and
repertoire_id names as headers.  Each value in the data frame represents the
frequency that the sequence appeared in the repertoire_id.
}
\seealso{
\code{\link{topSeqs}} and \code{\link{uniqueSeqs}}
}
