% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoringMatrix.R
\name{sorensenIndex}
\alias{sorensenIndex}
\title{Sorensen index}
\usage{
sorensenIndex(sample_list)
}
\arguments{
\item{sample_list}{A list of two tibbles consisting of frequencies of
antigen receptor sequences derived from the productiveSeq function
in LymphoSeq2.  "junction_aa", "duplicate_count" and 'repertoire_id'
are required columns.}
}
\value{
Returns the similarity score, a measure of the amount of
overlap between two samples.  The value ranges from 0 to 1 where 1 indicates
the sequence frequencies are identical in the two samples and 0
indicates no shared frequencies.
}
\description{
Calculates the Sorensen index between two groups of repertoires. Similar to
a Jaccard index, Sorensen index gives a greater weight to shared sequences
over unique sequences.
}
\examples{
file_path <- system.file("extdata", "TCRB_sequencing", package = "LymphoSeq2")

stable <- readImmunoSeq(path = file_path, threads = 1)

atable <- productiveSeq(stable, aggregate = "junction_aa")

}
\seealso{
\code{\link{scoringMatrix}}
}
