% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read10x.R
\name{standardize10x}
\alias{standardize10x}
\title{Standardize 10X data inputs}
\usage{
standardize10x(clone_file)
}
\arguments{
\item{clone_file}{A single file from 10x Genomics (AIRR .tsv or contig_annotations.csv)}
}
\value{
A tibble standardized to AIRR format.
}
\description{
Standardize 10X data inputs
}
