% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topSeqsPlot.R
\name{topSeqsPlot}
\alias{topSeqsPlot}
\title{Cumulative frequency bar plot of top sequences}
\usage{
topSeqsPlot(study_table, top = 10)
}
\arguments{
\item{study_table}{A study tibble imported using the LymphoSeq function readImmunoSeq
or productiveSeq.}

\item{top}{The number of top sequences to be colored in the bar plot.  All
other, less frequent sequences are colored violet.}
}
\value{
Returns a cumulative frequency bar plot of the top sequences.
}
\description{
Create a cumulative frequency bar plot of a specified number of top
sequences.
}
\details{
The plot is made using the package ggplot2 and can be reformatted
using ggplot2 functions.  See examples below.
}
\seealso{
An excellent resource for examples on how to reformat a ggplot can
be found in the R Graphics Cookbook online (\url{http://www.cookbook-r.com/Graphs/}).
}
