% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportFasta.R
\name{writeFasta}
\alias{writeFasta}
\title{Write FASTA file}
\usage{
writeFasta(study_table, type)
}
\arguments{
\item{study_table}{A tibble consisting of antigen receptor sequences
imported by the LymphoSeq function \code{\link[=readImmunoSeq]{readImmunoSeq()}}.}

\item{type}{A character vector indicating whether "junction_aa" or "junction"
sequences should be exported.  If "junction_aa" is specified, then run
\code{\link[=productiveSeq]{productiveSeq()}} first.}
}
\description{
Write FASTA file
}
