% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoringMatrix.R
\name{bhattacharyyaCoefficient}
\alias{bhattacharyyaCoefficient}
\title{Bhattacharyya coefficient}
\usage{
bhattacharyyaCoefficient(sample_list)
}
\arguments{
\item{sample_list}{A list of two tibble corresponding derived from the
\code{productiveSeq()}function in LymphoSeq2. "duplicate_frequency",
"junction_aa", and "repertoire_id" columns are necessary for the calculation
of the Bhattacharyya coefficient.}
}
\value{
A tibble with one row and three columns sample1, sample2,
bhattacharyya coefficient
}
\description{
Calculates the Bhattacharyya coefficient of two samples.
}
\seealso{
\code{\link[=scoringMatrix]{scoringMatrix()}}
}
