% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoringMatrix.R
\name{bhattacharyyaCoefficient}
\alias{bhattacharyyaCoefficient}
\title{Bhattacharyya coefficient}
\usage{
bhattacharyyaCoefficient(sample_list)
}
\arguments{
\item{sample_list}{A list of two tibble corresponding derived from the productiveSeq
function in LymphoSeq2. `duplicate_frequency`, `junction_aa`, and `repertoire_id` columns are necessary
for the calcualtion of the Bhattacharyya coefficient.}
}
\value{
A tibble with one row and three columns sample1, sample2, bhattacharyyaCoefficient
}
\description{
Calculates the Bhattacharyya coefficient of two samples.
}
\examples{
file_path <- system.file("extdata", "TCRB_sequencing", package = "LymphoSeq2")

stable <- readImmunoSeq(path = file_path)

atable <- productiveSeq(stable, aggregate = "junction_aa")

sample_list <- atable \%>\% 
               filter(repertoire_id \%in\% c("TRB_Unsorted_32", "TRB_Unsorted_83")) \%>\%
               group_by(repertoire_id) \%>\% group_split()

bhattacharyyaCoefficient(sample_list)
}
\seealso{
\code{\link{scoringMatrix}}
}
