% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chordDiagramVDJ.R
\name{chordDiagramVDJ}
\alias{chordDiagramVDJ}
\title{Chord diagram of VJ or DJ gene associations}
\usage{
chordDiagramVDJ(study_table, association = "VJ", colors = c("red", "blue"))
}
\arguments{
\item{study_table}{A tibble consisting of frequencies of antigen receptor
sequences.  "v_family", "j_family", and if applicable, "d_family"
are required columns.  Using output from the LymphoSeq2 function topSeqs is
recommended.}

\item{association}{A character vector of gene families to associate. Options
include "VJ" or "DJ".}

\item{colors}{A character vector of 2 colors corresponding to the V/D and J
gene colors respectively.}
}
\value{
Returns a chord diagram showing VJ or DJ gene associations from one or
more samples.
}
\description{
Creates a chord diagram showing VJ or DJ gene associations from one or more
samples.
}
\details{
The size of the ribbons connecting VJ or DJ genes correspond to the
number of samples or number of sequences that make up that recombination
event. The thicker the ribbon, the higher the frequency of the recombination.
}
\examples{
file_path <- system.file("extdata", "TCRB_sequencing", package = "LymphoSeq2")

stable <- readImmunoSeq(file_path)

ntable <- productiveSeq(stable, aggregate = "junction")

top_seqs <- topSeqs(ntable, top = 1)

chordDiagramVDJ(top_seqs, association = "VJ", colors = c("red", "blue"))


}
\seealso{
\code{\link{topSeqs}}
}
