% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonSeqsBar.R
\name{commonSeqsBar}
\alias{commonSeqsBar}
\title{Common sequences bar plot}
\usage{
commonSeqsBar(
  productive_aa,
  repertoire_ids,
  color_sample = NULL,
  color_intersection = NULL,
  color = "#377eb8",
  labels = "no"
)
}
\arguments{
\item{productive_aa}{A tibble of productive amino acid sequences
generated by LymphoSeq2 function productiveSeq where the aggregate parameter
was set to "junction_aa".}

\item{repertoire_ids}{The names of two or more repertoire_ids in the productive_aa
list whose intersections will shown.}

\item{color_sample}{The name of a single repertoire_id in the productive_aa list whose
sequences will be colored in all repertoire_ids that they appear in.}

\item{color_intersection}{The names of two or more repertoire_ids in the productive_aa
list whose intersections will be colored.}

\item{color}{A character vector of a color name that will be used highlight a selected
repertoire_id or multiple repertoire_id intersections.}

\item{labels}{A character vector indicating whether the number of
intersecting sequences should be shown on the tops of the bars.  Options
include "yes" or "no".}
}
\value{
Returns an UpSetR bar plot showing the number of intersecting sequences
across multiple repertoire_ids.
}
\description{
Creates an UpSetR bar plot showing the number of intersecting sequences across
multiple repertoire_ids. This function is useful when more than 3 repertoire_ids are being
compared.
}
\examples{
file_path <- system.file("extdata", "TCRB_sequencing", package = "LymphoSeq2")

stable <- readImmunoSeq(path = file_path)

atable <- productiveSeq(stable, aggregate = "junction_aa")

commonSeqsBar(atable, repertoire_ids = c("TRB_CD4_949", "TRB_CD8_949",
"TRB_Unsorted_949", "TRB_Unsorted_1320"), color_sample = "TRB_CD8_949")
}
\seealso{
\code{\link{productiveSeq}} \code{\link{commonSeqsVenn}} \code{\link{commonSeqsPlot}} \code{\link{commonSeqsBar}}
}
