% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonSeqsPlot.R
\name{commonSeqsPlot}
\alias{commonSeqsPlot}
\title{Common sequences scatter plot}
\usage{
commonSeqsPlot(sample1, sample2, amino_table, show = "common")
}
\arguments{
\item{sample1}{A name of a repertoire_id in a list of data frames generated by the
LymphoSeq2 function \code{\link[=productiveSeq]{productiveSeq()}}.}

\item{sample2}{A name of a repertoire_id in a list of data frames generated by the
LymphoSeq function \code{\link[=productiveSeq]{productiveSeq()}}.}

\item{amino_table}{A tibble of productive amino acid sequences produced by the
function \code{\link[=productiveSeq]{productiveSeq()}} containing the samples to be compared.}

\item{show}{A character vector specifying whether only the common sequences
should be shown or all sequences.  Available options are "common" or "all".}
}
\value{
Returns a frequency scatter plot of two samples showing only the
shared sequences.
}
\description{
Creates a scatter plot of just the sequences in common between two samples.
}
\details{
The plot is made using the package ggplot2 and can be reformatted
using ggplot2 functions. See examples below.
}
\examples{
file_path <- system.file("extdata", "TCRB_sequencing", package = "LymphoSeq2")
study_table <- readImmunoSeq(path = file_path)
study_table <- LymphoSeq2::topSeqs(study_table, top = 100)
amino_table <- productiveSeq(study_table = study_table, aggregate = "junction_aa")
commonSeqsPlot("TRB_Unsorted_32", "TRB_Unsorted_83",
  amino_table = amino_table
)
# Change the X and Y axis to log-10 scale
commonSeqsPlot("TRB_Unsorted_32", "TRB_Unsorted_83",
  amino_table = amino_table
) +
  ggplot2::scale_x_log10() +
  ggplot2::scale_y_log10() +
  ggplot2::annotation_logticks(sides = "bl")
}
\seealso{
An excellent resource for examples on how to reformat a ggplot can
be found in the R Graphics Cookbook online (\url{http://www.cookbook-r.com/Graphs/}).

\code{\link[=productiveSeq]{productiveSeq()}}, \code{\link[=commonSeqs]{commonSeqs()}},
\code{\link[=commonSeqsVenn]{commonSeqsVenn()}}, \code{\link[=commonSeqsBar]{commonSeqsBar()}}
}
