% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonSeqsVenn.R
\name{commonSeqsVenn}
\alias{commonSeqsVenn}
\title{Common sequences Venn diagram}
\usage{
commonSeqsVenn(repertoire_ids, productive_aa)
}
\arguments{
\item{repertoire_ids}{A character vector of two or three names of repertoire_ids in 
productiveSeq table to compare.}

\item{productive_aa}{A tibble of amino acid sequences generated
by the LymphoSeq function productiveSeq.}
}
\value{
Returns a a Venn diagram of the number of common sequences between
two or three repertoire_ids.
}
\description{
Creates a Venn diagram comparing the number of common sequences in two or 
three repertoire_ids.
}
\examples{
file.path <- system.file("extdata", "TCRB_sequencing", package = "LymphoSeq2")

study_table <- readImmunoSeq(path = file.path)

productive_aa <- productiveSeq(study_table = study_table, aggregate = "junction_aa")

# Plot a triple Venn diagram
commonSeqsVenn(repertoire_ids = c("TRB_Unsorted_0", 
   "TRB_Unsorted_32", "TRB_Unsorted_83"), 
   productive_aa = productive_aa)

# Plot a double Venn diagram
commonSeqsVenn(repertoire_ids = c("TRB_Unsorted_0", 
   "TRB_Unsorted_32"), productive_aa = productive_aa)

}
\seealso{
\code{\link{commonSeqs}}
}
