% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countKmer.R
\name{countKmer}
\alias{countKmer}
\title{Find kmers and its counts}
\usage{
countKmer(study_table, k, separate = TRUE)
}
\arguments{
\item{study_table}{A tibble outputed from the readImmunoSeq function.}

\item{k}{The length of kmers.}

\item{separate}{A boolean value. TRUE to separate the counts of each k-mer
by repertoire_ids. FALSE to show cumulative counts instead.}
}
\value{
A tibble with the k-mer and its counts. The counts can be cumulative
counts of entire study_table or counts by repertoire_id.
}
\description{
Calculate counts of kmers in the query nucleotide sequence
}
\examples{
file_path <- system.file("extdata", "TCRB_sequencing", package = "LymphoSeq2")

stable <- readImmunoSeq(path = file_path)

kmer_table <- countKmer(study_table = stable, k = 5, separate = TRUE)

}
