% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countKmer.R
\name{countKmer}
\alias{countKmer}
\title{Find kmers and its counts}
\usage{
countKmer(study_table, k)
}
\arguments{
\item{study_table}{A tibble outputed from the readImmunoSeq function.}

\item{k}{The length of kmers.}
}
\value{
A tibble with two columns: the kmers and its respective count.
}
\description{
Calculate counts of kmers in the query nucleotide sequence
}
\examples{
file_path <- system.file("extdata", "TCRB_sequencing", package = "LymphoSeq2")

stable <- readImmunoSeq(path = file_path)

countKmer(study_table = stable, k = 5)

}
