% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportFasta.R
\name{exportFasta}
\alias{exportFasta}
\title{Export sequences in fasta format}
\usage{
exportFasta(
  study_table,
  type = "junction",
  names = c("rank", "junction_aa", "duplicate_count")
)
}
\arguments{
\item{type}{A character vector indicating whether "junction_aa" or "junction" sequences
should be exported.  If "junction_aa" is specified, then run productiveSeqs first.}

\item{names}{A character vector of one or more column names to name the sequences.
If "rank" is specified, then the rank order of the sequences by frequency is used.}

\item{sample_table}{A tibble consisting of antigen receptor sequences 
imported by the LymphoSeq function readImmunoSeq.}
}
\value{
Exports fasta files to the working directory.
}
\description{
Export junction or amino acid sequences in fasta format.
}
\examples{
file.path <- system.file("extdata", "TCRB_sequencing", package = "LymphoSeq")

study_table <- readImmunoSeq(path = file.path)

exportFasta(study_table = study_table, type = "junction", names = c("rank", "junction_aa", "duplicate_count"))

productive_aa <- productiveSeq(study_table = study_table, aggregate = "junction_aa")

exportFasta(list = productive_aa, type = "junction_aa", names = "duplicate_frequency")
}
