% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialAbundance.R
\name{fisherFunction}
\alias{fisherFunction}
\title{Fisher test}
\usage{
fisherFunction(x, y, not_x, not_y)
}
\arguments{
\item{x}{Number of instances of group 1}

\item{y}{Number of instances of group 2}

\item{not_x}{Number of instances not belonging to group 1}

\item{not_y}{Number of instances not belonging to group 2}
}
\description{
Fisher test
}
