% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneFreq.R
\name{geneFreq}
\alias{geneFreq}
\title{Gene frequencies}
\usage{
geneFreq(productive_nt, locus = "V|D|J", family = FALSE)
}
\arguments{
\item{locus}{A character vector indicating which VDJ genes to include in the 
output.  Available options include "VDJ", "DJ", "VJ", "DJ", "V", "D", or "J".}

\item{family}{A Boolean value indicating whether or not family names instead 
of gene names are used.  If TRUE, then family names are used and if FALSE, 
gene names are used.}

\item{productive_table}{A tibble of productive sequences 
generated by the LymphoSeq function productiveSeq where the parameter 
aggregate is set to "nucleotide".}
}
\value{
Returns a data frame with the repertoire_id names, VDJ gene name, duplicate_count, and 
\% frequency of the V, D, or J genes (each gene frequency should add to 
100\% for each repertoire_id).
}
\description{
Creates a data frame of VDJ gene counts and frequencies.
}
\examples{
file_path <- system.file("extdata", "TCRB_sequencing", package = "LymphoSeq")

stable <- readImmunoSeq(path = file_path)

ntable <- productiveSeq(study_table = stable, aggregate = "nucleotide")

geneFreq(ntable, locus = "VDJ", family = FALSE)

}
