% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonalRelatedness.R
\name{getRelatedness}
\alias{getRelatedness}
\title{Calculate relatedness}
\usage{
getRelatedness(sample_table, editDistance = 10)
}
\arguments{
\item{sample_table}{A tibble of unproductive or productive junction
sequences or productive junction sequences.  junction and duplicate_count are
required columns.}

\item{editDistance}{An integer giving the minimum edit distance that the
sequence must be less than or equal to. See details below.}
}
\description{
Calculates the clonal relatedness of a repertoire_id.
}
