% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonalRelatedness.R
\name{getRelatedness}
\alias{getRelatedness}
\title{Calculate relatedness}
\usage{
getRelatedness(study_table, edit_distance = 10)
}
\arguments{
\item{study_table}{A tibble of raw or productive nucleotide junction
sequences."junction" and "duplicate_count" are required columns.}

\item{edit_distance}{An integer giving the minimum edit distance that the
sequence must be less than or equal to. See details below. Deafault: 10}
}
\description{
Calculates the clonal relatedness of a repertoire.
}
