% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonality.R
\name{iterativeSummary}
\alias{iterativeSummary}
\title{Uncount, sample and iteratively calculate repertoire summary statistics}
\usage{
iterativeSummary(study_table, iterations, min_count = 1000)
}
\arguments{
\item{study_table}{A tibble consisting of antigen receptor sequencing
imported by the LymphoSeq2 function \code{\link[=readImmunoSeq]{readImmunoSeq()}}. "junction_aa",
"duplicate_count", and "duplicate_frequency" are required columns. Note that
clonality is usually calculated from productive junction sequences.
Therefore, it is not recommended to run this function using a productive
sequence list aggregated by amino acids.}

\item{iterations}{Number of iterations to run the sampled clonality metrics.}

\item{min_count}{The minimum depth to which each repertoire in the study must
be sampled to. Default 1000 (the default)}
}
\value{
Tibble summarizing the sequence information for each repertoire_id
normalized by depth of sequencing
}
\description{
Uncount, sample and iteratively calculate repertoire summary statistics
}
