% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countKmer.R
\name{kmerPlot}
\alias{kmerPlot}
\title{Plot kmer distributions}
\usage{
kmerPlot(kmer_table, top = 10)
}
\arguments{
\item{kmer_table}{A tibble output from the countKmer function.}

\item{top}{The number of top kmers to show}
}
\value{
A stacked bar chart showing kmer distributions
}
\description{
Plot kmer distributions by repertoire id
}
\examples{
file_path <- system.file("extdata", "TCRB_sequencing", package = "LymphoSeq2")

stable <- readImmunoSeq(path = file_path)

kmer_table <- countKmer(study_table = stable, k = 5, separate = TRUE)

kmer_distributions <- kmerPlot(kmer_table, top = 10)

}
