% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LymphoSeq2-package.R
\docType{package}
\name{LymphoSeq2-package}
\alias{LymphoSeq2}
\alias{LymphoSeq2-package}
\title{LymphoSeq2: Analyze High-Throughput Sequencing of T and B Cell Receptors}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Analyze high-throughput sequencing of T and B cell receptor complementarity determining region 3 (CDR3) sequences. Supports multiple platforms including Adaptive Biotechnologies ImmunoSEQ, 10X Genomics single-cell VDJ sequencing, BGI-IRSeq, and MiXCR processed data. Provides tools for diversity analysis, clonality metrics, gene usage visualization, and repertoire comparison.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shashidhar22/LymphoSeq2}
  \item \url{https://shashidhar22.github.io/LymphoSeq2}
  \item Report bugs at \url{https://github.com/shashidhar22/LymphoSeq2/issues}
}

}
\author{
\strong{Maintainer}: Ravishankar Shashidhar \email{shashidhar.r.shankar@gmail.com}

Authors:
\itemize{
  \item Wu Elena \email{ewu2@fredhutch.org}
  \item Coffey David
}

}
\keyword{internal}
