% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwisePlot.R
\name{pairwisePlot}
\alias{pairwisePlot}
\title{Pairwise comparison plot}
\usage{
pairwisePlot(matrix)
}
\arguments{
\item{matrix}{A similarity or Bhattacharyya matrix produced by the LymphoSeq2 
scoringMatrix function.}
}
\value{
A pairwise comparison heat map.
}
\description{
Creates a heat map from a similarity or Bhattacharyya matrix.
}
\details{
The plot is made using the package ggplot2 and can be reformatted
using ggplot2 functions.  See examples below.
}
\examples{
file.path <- system.file("extdata", "TCRB_sequencing", package = "LymphoSeq")

study_table <- readImmunoSeq(path = file.path)

productive_aa <- productiveSeq(study_table = study_table, aggregate = "aminoAcid")

similarity_matrix <- scoringMatrix(productive_table = productive_aa, mode="Similarity")

pairwisePlot(matrix = similarity_matrix)

bhattacharyya_matrix <- scoringMatrix(productive_table = productive_aa, mode="Bhattacharyya")

pairwisePlot(matrix = bhattacharyya_matrix)

# Change plot color, title legend, and add title
pairwisePlot(matrix = similarity.matrix) + 
   ggplot2::scale_fill_gradient(low = "#deebf7", high = "#3182bd") + 
   ggplot2::labs(fill = "Similarity score") + ggplot2::ggtitle("Figure Title")
}
\seealso{
An excellent resource for examples on how to reformat a ggplot can 
be found in the R Graphics Cookbook online (\url{http://www.cookbook-r.com/Graphs/}).
The functions to create the similarity or Bhattacharyya matrix can be found 
here: \code{\link{similarityMatrix}} and \code{\link{bhattacharyyaMatrix}}
}
