% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoringMatrix.R
\name{percentSI}
\alias{percentSI}
\title{Percent similarity index}
\usage{
percentSI(sample_list)
}
\arguments{
\item{sample_list}{A list of two tibble corresponding derived from the
\code{\link[=productiveSeq]{productiveSeq()}} function in LymphoSeq2. "duplicate_frequency",
"junction_aa", and "repertoire_id" columns are necessary for the calculation
of the Bhattacharyya coefficient.}
}
\value{
Returns the similarity score, a measure of the amount of
overlap between two samples.  The value ranges from 0 to 1 where 1 indicates
the sequence frequencies are identical in the two samples and 0
indicates no shared frequencies.
}
\description{
Calculates the Percent similarity index between two groups of repertoires.
Percent similarity index, not only compares the number of similar and
dissimilar species present between two sites, but also incorporate abundance.
}
\seealso{
\code{\link[=scoringMatrix]{scoringMatrix()}}
}
