% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloTree.R
\name{phyloTree}
\alias{phyloTree}
\title{Create phylogenetic tree}
\usage{
phyloTree(
  study_table,
  repertoire_ids,
  type = "junction",
  layout = "rectangular",
  label = TRUE
)
}
\arguments{
\item{study_table}{A tibble of unproductive junction sequences or productive 
junction sequences generated by the LymphoSeq function productiveSeq.  
v_family, d_family, j_family, and duplicate_count are required columns.}

\item{repertoire_ids}{A character vector indicating the name of the repertoire_id in the study table.}

\item{type}{A character vector indicating whether "junction_aa" or "junction" sequences
should be compared.}

\item{layout}{A character vector indicating the tree layout.  Options include 
"rectangular", "slanted", "fan", "circular", "radial" and "unrooted".}

\item{label}{A Boolean indicating if the sequencing duplicate_count should be shown next to the leaves.}
}
\value{
Returns a phylogenetic tree where each leaf represents a sequence color coded by the
V, D, and J gene usage.  The number next to each leaf refers to the sequence duplicate_count.  A triangle 
shaped leaf indicates the dominant sequence.  Refer to the ggtree Bioconductor package 
documentation for details on how to manipulate the tree.
}
\description{
Create a phylogenetic tree using neighbor joining tree estimation for amino acid or 
junction CDR3 sequences in a list of data frames.
}
\examples{
file_path <- system.file("extdata", "IGH_sequencing", package = "LymphoSeq2")

stable <- readImmunoSeq(path = file_path)

ntable <- productiveSeq(study_table = stable, aggregate = "junction")

phyloTree(study_table = ntable, repertoire_id = "IGH_MVQ92552A_BL", type = "junction", 
         layout = "rectangular")

phyloTree(study_table = ntable, repertoire_id = "IGH_MVQ92552A_BL", type = "junction_aa", 
         layout = "circular")
         
# Add scale and title to figure
phyloTree(study_table = ntable, repertoire_id = "IGH_MVQ92552A_BL", type = "junction_aa", 
         layout = "rectangular") +
         ggtree::theme_tree2() +
         ggplot2::theme(legend.position = "right", legend.key = element_rect(colour = "white")) +
         ggplot2::ggtitle("Title")
         
# Hide legend and leaf labels
phyloTree(study_table = productive_nt, repertoire_id = "IGH_MVQ92552A_BL", type = "junction", 
         layout = "rectangular", label = FALSE) +
         ggplot2::theme(legend.position="none")
}
