% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefactionCurve.R
\name{plotRarefactionCurve}
\alias{plotRarefactionCurve}
\title{Plot rarefication and extrapolation curves for samples}
\usage{
plotRarefactionCurve(study_table)
}
\arguments{
\item{study_table}{A tibble consisting antigen receptor sequencing
data imported by the LymphoSeq function readImmunoSeq. "aminoAcid", "count", and
"frequencyCount" are required columns.}
}
\description{
Given a study table, for each sample plot rarefaction curves to estimate
repertoire diversity. The method used to generate the rarefaction curve
is derived from Chao et al., (2014) using the iNEXT library
}
\examples{
file_path <- system.file("extdata", "TCRB_sequencing", package = "LymphoSeq2")
stable <- readImmunoSeq(path = file_path)


}
