% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrackSingular.R
\name{plotTrackSingular}
\alias{plotTrackSingular}
\alias{highlightPlot}
\title{Highlight clones}
\usage{
plotTrackSingular(clone_table)

highlightPlot(aseq, clone_table)
}
\arguments{
\item{clone_table}{tibble of productive amino acid sequences to highlight
generated by LymphoSeq function cloneTrack}

\item{aseq}{CDR3 amino acid sequence to highlight}
}
\value{
A list of alluvial plots highlighting a single sequence
}
\description{
Create alluvial plots highlighting each sequence in amino acid list
}
\details{
The plot is made using the package ggplot2 and can be reformatted
using ggplot2 functions.
}
\section{Functions}{
\itemize{
\item \code{highlightPlot}: Highlight a specific amino acid sequence
}}

\examples{
file_path <- system.file("extdata", "TCRB_sequencing", package = "LymphoSeq2")
study_table <- LymphoSeq2::readImmunoSeq(path = file_path)
study_table <- LymphoSeq2::topSeqs(study_table, top = 100)
amino_table <- LymphoSeq2::productiveSeq(study_table, aggregate = "junction_aa")
top_seq <- LymphoSeq2::topSeqs(amino_table, top = 3)
clone_table <- LymphoSeq2::cloneTrack(
  study_table = top_seq,
  sample_list = c("TRB_CD8_949", "TRB_CD8_CMV_369")
)
LymphoSeq2::plotTrackSingular(clone_table)
}
