% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prevalenceTRB}
\alias{prevalenceTRB}
\title{Prevalence TRB}
\format{
\subsection{\code{prevalenceTRB}}{

A tibble with 11,724,294 rows and 2 columns:
\describe{
\item{prevalence}{Frequency of sequences in 55 healthy PBMC samples}
\item{aminoAcid}{T-cell receptor beta chain amino acid sequence}
}
}
}
\source{
TCR beta sequencing data from PBMCs of 55 healthy individuals
}
\usage{
prevalenceTRB
}
\description{
The database describes the frequency at which a CDR3 amino acid sequence was
found in cohort of 55 PBMC samples from healthy individuals
}
\keyword{datasets}
