% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read10x.R
\name{read10x}
\alias{read10x}
\title{Read 10x Genomics files}
\usage{
read10x(path, recursive = FALSE)
}
\arguments{
\item{path}{Path to the directory containing .tsv files. Only
files with the .tsv extension are imported.}

\item{recursive}{A Boolean value
\itemize{
\item \code{TRUE} : the function will recursively search directory for all .tsv files
\item \code{FALSE} (the default): Open file using path
}}
}
\value{
Returns a tibble with MiAIRR headers and repertoire_id
}
\description{
Imports tab-separated value (.tsv) files exported by 10x Genomics
and stores them as MiAIRR compliant tibbles.
}
