% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read10x.R
\name{read10x}
\alias{read10x}
\title{Read 10x Genomics files}
\usage{
read10x(clone_data)
}
\arguments{
\item{clone_data}{An MiAIRR compliant tibble generated by readImmunoSeq}
}
\value{
Returns a tibble with MiAIRR headers and repertoire_id
}
\description{
Imports tab-separated value (.tsv) files exported by 10x Genomics
and stores them as MiAIRR compliant tibbles.
}
