% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readImmunoSeq.R
\name{readImmunoSeq}
\alias{readImmunoSeq}
\alias{getStandard}
\alias{readFiles}
\title{Read ImmunoSeq files}
\usage{
readImmunoSeq(path, recursive = FALSE)

getStandard(file_type)

readFiles(clone_file, progress)
}
\arguments{
\item{path}{Path to AIRR-seq files. Can be a file path, directory path or a vector of file paths}

\item{recursive}{Recursively search directory for AIRR-seq files. Boolean value}

\item{file_type}{file type from getFileType method}

\item{clone_file}{.tsv file containing results from AIRRSeq pipeline}
}
\value{
Returns a tibble with MiAIRR headers and repertoire_id

List of standard header names

Tibble in MiAIRR format
}
\description{
\code{readImmunoSeq} Imports tab-separated value (.tsv) files exported by the Adaptive 
Biotechnologies ImmunoSEQ analyzer, BGI IR-SEQ, MiXCR and stores 
them as MiAIRR compliant tibble.
}
\details{
May import tab-delimited files containing antigen receptor 
sequencing from with the following header set.  

| MiAIRR field ID | Type1           | Type2                       | Type3               | Type4                          | Type5                               | Type6             |
| --------------- | --------------- | --------------------------- | ------------------- | ------------------------------ | ----------------------------------- | ----------------- |
| junction        | nucleotide      | nucleotide                  | nucleotide          | nucleotide.CDR3.in.lowercase . | nucleotide( CDR3   in   lowercase ) | nSeqCDR3          |
| junction_aa     | aminoAcid       | aminoAcid                   | aminoAcid           | aminoAcid.CDR3.in.lowercase .  | aminoAcid( CDR3   in   lowercase )  | aaSeqCDR3         |
| duplicate_count | count ( reads ) | count ( templates / reads ) | count ( templates ) | cloneCount                     | cloneCount                          | cloneCount        |
| v_call          | vGeneName       | vGeneName                   | vGeneName           | vGene                          | vGene                               | allVHitsWithScore |
| d_call          | dGeneName       | dGeneName                   | dGeneName           | dGene                          | dGene                               | allDHitsWithScore |
| j_call          | jGeneName       | jGeneName                   | jGeneName           | jGene                          | jGene                               | allJHitsWithScore |
}
\section{Get standard headers}{


Retrives MiAIRR standard compliant fields from the clone files
}

\section{Read clone file from path}{


Given the path to a single AIRRSeq clone file, generate MiAIRR compliant tibble.
}

\examples{
file_path <- base::system.file("extdata", "TCRB_study", package = "LymphoSeq2")
study_table <- LymphoSeq2::readImmunoSeq(path = file_path, 
                                         recursive = FALSE)
                             
file_list <- base::list.files(file_path, 
                              full.name = TRUE,
                              all.files = FALSE,
                              pattern = ".tsv",
                              include.dirs = FALSE)
study_table <- LymphoSeq2::readImmunoSeq(path = file_list[1],
                                         recursive = FALSE)
study_table <- LymphoSeq2::readImmunoSeq(path = file_list[1:2],
                                         recursive = FALSE)
}
