% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readImmunoSeq.R
\name{readImmunoSeq}
\alias{readImmunoSeq}
\title{Read ImmunoSeq files}
\usage{
readImmunoSeq(path, recursive = FALSE, threads = parallel::detectCores()/2)
}
\arguments{
\item{path}{Path to the directory containing tab-delimited files. Only files
with the extension .tsv are imported. The names of the data frames are
the same as names of the files.}

\item{recursive}{A Boolean value
\itemize{
\item \code{TRUE} : the function will recursively search directory for all .tsv files
\item \code{FALSE} (the default): Open file using path
}}

\item{threads}{Number of threads.}
}
\value{
Returns a tibble with MiAIRR headers and repertoire_id
}
\description{
\code{\link[=readImmunoSeq]{readImmunoSeq()}} Imports tab-separated value (.tsv) files exported by the
Adaptive Biotechnologies ImmunoSEQ analyzer, BGI IR-SEQ, MiXCR and stores
them as MiAIRR compliant tibble.
}
\examples{
file_path <- system.file("extdata", "TCRB_sequencing", package = "LymphoSeq2")
study_table <- LymphoSeq2::readImmunoSeq(path = file_path, recursive = FALSE,
  threads = 1)
study_table <- LymphoSeq2::topSeqs(study_table, top = 100)

}
