% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeSeq.R
\name{removeSeq}
\alias{removeSeq}
\title{Remove sequence}
\usage{
removeSeq(study_table, sequence)
}
\arguments{
\item{sequence}{A character vector of one or more amino acid sequences to 
remove from the list of data frames.}

\item{file.list}{A list of data frames imported using the LymphoSeq function 
readImmunoSeq.  "aminoAcid", "count", and "frequencyCount" are required columns.}
}
\value{
Returns a list of data frames like the one imported except all rows 
with the specified amino acid sequence are removed.  The frequencyCount is 
recalculated.
}
\description{
Removes an amino acid sequence and associated data from all instances within 
a list of data frames and then recomputes the frequencyCount.
}
\examples{
file_path <- system.file("extdata", "TCRB_sequencing", package = "LymphoSeq2")
stable <- readImmunoSeq(path = file_path)
searchSeq(stable, sequence = "CASSDLIGNGKLFF")
cleansed <- removeSeq(stable, sequence = "CASSDLIGNGKLFF")
searchSeq(cleansed, sequence = "CASSDLIGNGKLFF")
}
