% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runINext.R
\name{runINext}
\alias{runINext}
\title{Run iNEXT on samples}
\usage{
runINext(sample_table, progress, color = "sample")
}
\arguments{
\item{sample_table}{A tibble consisting antigen receptor sequencing 
data imported by the LymphoSeq function readImmunoSeq. "aminoAcid", "count", and 
"frequencyCount" are required columns.}
}
\description{
Given a sample table, for each generate rarefaction curves to estimate 
repertoire diversity. The method used to generate the rarefaction curve
is derived from Chao et al., (2014) using the iNEXT library
}
\examples{
file.path <- system.file("extdata", "TCRB_sequencing", package = "LymphoSeq")

study_table <- readImmunoSeq(path = file.path)

productive_aa <- productiveSeq(study_table = study_table, 
  aggregate = "aminoAcid", prevalence = TRUE)

sample_table <- productive_aa \%>\% filter(sample == "TRB_Unsorted_1320")

}
