% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonality.R
\name{sampledSummary}
\alias{sampledSummary}
\title{Sample repertoire}
\usage{
sampledSummary(study_table, min_count)
}
\arguments{
\item{study_table}{A tibble consisting of antigen receptor sequencing
imported by the LymphoSeq2 function \code{\link[=readImmunoSeq]{readImmunoSeq()}}. "junction_aa",
"duplicate_count", and "duplicate_frequency" are required columns. Note that
clonality is usually calculated from productive junction sequences.
Therefore, it is not recommended to run this function using a productive
sequence list aggregated by amino acids.}

\item{min_count}{The minimum depth to which each repertoire in the study must
be sampled to. Default 1000 (the default)}
}
\value{
Tibble summarizing the sequence information for each repertoire_id
randomly sampled to the minimum count
}
\description{
Sample repertoire
}
