% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchDB.R
\name{searchDB}
\alias{searchDB}
\title{Search for T cell receptor beta CDR3 amino acid sequences with known antigen
specificity from iReceptor2}
\usage{
searchDB(study_table, credential)
}
\arguments{
\item{study_table}{A tibble generated by the LymphoSeq functions
searchPublished or topSeqs.  "aminoAcid", "frequencyCount", and "count"
are required columns.}
}
\value{
Returns a tibble of with sample name and instance in the sample
that the published TCR sequence appeared along with additional
information including disease status, sample type, and PubMed ID
(PMID) for the reference where the sequence was characterized.
}
\description{
Search for published T cell receptor beta CDR3 amino acid sequences with
known antigen specificity in a list of data frames.
}
