% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchDB.R
\name{searchDB}
\alias{searchDB}
\title{Search for T cell receptor beta CDR3 amino acid sequences with known antigen
specificity from IEDB, VdjDB and McPASDb}
\usage{
searchDB(study_table, dbname = "all", chain = "trb")
}
\arguments{
\item{study_table}{A tibble generated by the LymphoSeq2 functions
\code{\link[=readImmunoSeq]{readImmunoSeq()}}, \code{\link[=productiveSeq]{productiveSeq()}}, \code{\link[=searchPublished]{searchPublished()}} or \code{\link[=topSeqs]{topSeqs()}}.
"junction_aa", "duplicate_frequency", and "duplicate_count"
are required columns.}

\item{dbname}{A vector of database source to search for the antigenic
specificity of a T-cell or B-cell in the dataset: \code{"all"}, \code{"IEDB"},
\code{"McPAS-TCR"}, \code{"VdjDB"}.}

\item{chain}{The receptor chain type to search in the public databases:
\code{"tra"}, \code{"trb"}, \code{"light"}, \code{"heavy"}.}
}
\value{
Returns the input table annotated with the any recorded antigenic
specificity from the public databases.
}
\description{
Search for published T cell receptor beta CDR3 amino acid sequences with
known antigen specificity in a list of data frames.
}
\examples{
file_path <- system.file("extdata", "TCRB_sequencing", package = "LymphoSeq2")
study_table <- LymphoSeq2::readImmunoSeq(path = file_path)
study_table <- LymphoSeq2::topSeqs(study_table, top = 100)
amino_table <- LymphoSeq2::productiveSeq(study_table = study_table, aggregate = "junction_aa")
top_seqs <- LymphoSeq2::topSeqs(productive_table = amino_table, top = 1)
LymphoSeq2::searchDB(study_table = top_seqs, dbname = "all", chain = "trb")
}
