% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read10x.R
\name{selectChain}
\alias{selectChain}
\title{Select the most frequent chain}
\usage{
selectChain(barcode_data, clone_data, chain = "TRA")
}
\arguments{
\item{barcode_data}{A tibble that holds data for one barcode
identifier}

\item{clone_data}{A tibble that holds data for one repertoire_id
(one file)}

\item{chain}{The chain to examine to select the most frequently
occurring one. Values given can only be "TRA" or "TRB" to indicate
alpha or beta chain respectively.}
}
\value{
A tibble with one row of data that contains the most
frequently occurring chain.
}
\description{
Select the most frequent chain
}
