% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoringMatrix.R
\name{similarityScore}
\alias{similarityScore}
\title{Similarity score}
\usage{
similarityScore(sample_list)
}
\arguments{
\item{sample1}{A data frame consisting of frequencies of antigen receptor 
sequences.  "junction_aa" and "duplicate_count" are a required columns.}

\item{sample2}{A data frame consisting of frequencies of antigen receptor 
sequences.  "junction_aa" and "duplicate_count" are a required columns.}
}
\value{
Returns the similarity score, a measure of the amount of 
overlap between two samples.  The value ranges from 0 to 1 where 1 indicates 
the sequence frequencies are identical in the two samples and 0 
indicates no shared frequencies.
}
\description{
Calculates the similarity score of two samples.
}
\examples{
file.path <- system.file("extdata", "TCRB_sequencing", package = "LymphoSeq")

study_table <- readImmunoSeq(path = file.path)

productive.aa <- productiveSeq(study_table, aggregate = "junction_aa")

sample_list <- productive.aa \%>\% 
               filter(repertoire_id \%in\% c("TRB_Unsorted_32", "TRB_Unsorted_83")) \%>\%
               group_by(repertoire_id) \%>\% group_split()

similarityScore(sample_list)
}
\seealso{
\code{\link{scoringMatrix}}
}
