% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoringMatrix.R
\name{sorensenIndex}
\alias{sorensenIndex}
\title{Sorensen index}
\usage{
sorensenIndex(sample_list)
}
\arguments{
\item{sample1}{A data frame consisting of frequencies of antigen receptor
sequences.  "junction_aa" and "duplicate_count" are a required columns.}

\item{sample2}{A data frame consisting of frequencies of antigen receptor
sequences.  "junction_aa" and "duplicate_count" are a required columns.}
}
\value{
Returns the similarity score, a measure of the amount of
overlap between two samples.  The value ranges from 0 to 1 where 1 indicates
the sequence frequencies are identical in the two samples and 0
indicates no shared frequencies.
}
\description{
Calculates the Sorensen index between two groups of repertoires. Similar to
a Jaccard index, Sorensen index gives a greater weight to shared sequences
over unique sequences.
}
\seealso{
\code{\link{scoringMatrix}}
}
