% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonality.R
\name{summarySeq}
\alias{summarySeq}
\title{Get summary statistics for each repertoire_id in the analysis}
\usage{
summarySeq(study_table)
}
\arguments{
\item{sample_table}{immune repertoire tibble for each a repertoire_id}
}
\value{
Tibble summarizing the sequence infomration for each repertoire_id
}
\description{
Get summary statistics for each repertoire_id in the analysis
}
