% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueSeqs.R
\name{uniqueSeqs}
\alias{uniqueSeqs}
\title{Unique sequences}
\usage{
uniqueSeqs(productive_table = productive_table, unique_type = "junction_aa")
}
\arguments{
\item{productive_table}{A tibble of productive amino acid sequences
imported using the function LymphoSeq2 function \code{\link[=productiveSeq]{productiveSeq()}} where the
aggregate parameter was set to "junction_aa".}

\item{unique_type}{Use \code{"junction_aa"} (the default) to aggregate by amino
acid sequences. Use \code{"junction"} to aggregate by nucleotide sequences.}
}
\value{
A data frame of unique amino acid sequences from the list of
data frames aggregated by "duplicate_count".
}
\description{
Aggregates all productive sequences within a list of data frames by
duplicate_count.
}
\examples{
file_path <- system.file("extdata", "TCRB_sequencing",
 package = "LymphoSeq2")
study_table <- LymphoSeq2::readImmunoSeq(path = file_path, threads = 1)
study_table <- LymphoSeq2::topSeqs(study_table, top = 100)
amino_table <- LymphoSeq2::productiveSeq(study_table = study_table,
 aggregate = "junction_aa")
unique_seqs <- LymphoSeq2::uniqueSeqs(productive_table = amino_table,
 unique_type = "junction_aa")
}
